requirements/requirementsd.txt
alabaster==0.7.11
asn1crypto==0.24.0
autopep8==1.4
Babel==2.6.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.1
cryptography==2.3.1
decorator==4.3.0
defusedxml==0.5.0
diff-match-patch==20121119
Django==2.1.1
django-admin-rangefilter==0.3.7
django-chartjs==1.3
django-decouple==2.1
django-jet==1.0.7
django-otp==0.5.0
django-rest-swagger==2.2.0
django-secrets==1.0.2
django-silk==3.0.1
django-sslserver==0.20
djangorestframework==3.8.2
djangorestframework-xml==1.3.0
docutils==0.14
et-xmlfile==1.0.1
feedparser==5.2.1
flake8==3.5.0
future==0.16.0
get==1.0.3
gprof2dot==2016.10.13
idna==2.7
imagesize==1.1.0
itypes==1.1.0
jdcal==1.4
Jinja2==2.10
MarkupSafe==1.0
mccabe==0.6.1
odfpy==1.3.6
openapi-codec==1.3.2
openpyxl==2.5.8
packaging==17.1
pluggy==0.7.1
post==1.0.2
psycopg2==2.7.5
psycopg2-binary==2.7.5
public==1.0.3
py==1.6.0
pycodestyle==2.3.1
pycparser==2.19
pyflakes==1.6.0
Pygments==2.2.0
pyparsing==2.2.1
python-dateutil==2.7.3
pytz==2018.5
PyYAML==3.13
query-string==1.0.2
request==1.0.2
requests==2.19.1
simplejson==3.16.0
six==1.11.0
snowballstemmer==1.2.1
Sphinx==1.8.1
sphinxcontrib-websupport==1.1.0
sqlparse==0.2.4
tablib==0.12.1
toml==0.9.6
tox==3.4.0
unicodecsv==0.14.1
uritemplate==3.0.0
urllib3==1.23
virtualenv==16.0.0
xlrd==1.1.0
xlwt==1.3.0
